.dir-locals.el: Prevent errors if .dir-locals.el isn't found.
While this repo should presumably always have a .dir-locals.el
by the nature of this file itself, it seems that this behavior "leaks".
See added comment for more details on the fix to this strange bug,
which is likely an upstream emacs or vc-mode issue.

Thanks to Miguel Ángel Arruga Vivas and Maxim Cournoyer for helping
investigate this problem.

* .dir-locals.el: Don't error out if .dir-locals.el isn't found.