gnu: python-pillow: Add the python2-variant property.
This is apparently required to get the package-with-python2 package
transformer to pick up the correct python2-pillow package instead of
attempting to create a new one with the same version as python-pillow, as it
would happen for python2-matplotlib for example.

* gnu/packages/python-xyz.scm (python-pillow): Add python2-variant property.
(python2-pillow): Strip python2-variant property from inherited python-pillow,
and fix indentation.