gnu: emacs-yasnippet-snippets: Resolve void function error.
Attempting to use the python-mode snippets shipped with this package would
result in a '(void-function yasnippet-snippets--fixed-indent)' error being
raised (failing the initialization of other packages such as Elpy).  The
following change fixes this by not only installing the snippets but also the
accompanying Elisp library that allows automatic discovery of the snippets.

* gnu/packages/emacs-xyz.scm (emacs-yasnippet-snippets)[build-system]: Switch
to emacs-build-system.
[arguments]: Remove the #:module and #:builder arguments.  Add the #:include
argument.
[description]: Adjust to reflect that it is no longer necessary to manually
add the snippets to the yas-snippet-dirs variable.