gnu: MariaDB: Fix 'mariadb_config' output to refer to correct directories.
This fixes a regression introduced in ce29031a23104928155b498d552fc4140cc03030
where 'mariadb_config' would return libdir and plugindir relative to the
location of the 'mariadb_config' program (which lives in a separate output).

While at it, fix a couple other wrong references, and a another regression
related to test installation.

* gnu/packages/databases.scm (mariadb)[arguments]: Patch "mariadb_config.c.in"
so that it falls back to old behavior.  Patch 'mariadb.pc.in' and
'mariadb_connector_c.cmake' to cope with absolute directory names. In
 #:configure-flags, don't set INSTALL_MYSQLTESTDIR, but delete the installed
tests with a phase.