gnu: libffi: On powerpc, patch with --force.
As explained in bug 45252, this is the right way to invoke it.

Efraim has confirmed that it is OK to make this change on master branch for
the generic powerpc case, even though it will in theory cause rebuilds on that
architecture, since in practice it isn't actually being used yet.

* gnu/packages/libffi.scm (arguments): Change the invocation of the "patch"
tool in the case where the %current-target-system or %current-system begins
with "powerpc", so that we invoke it using --force instead of --batch.