gnu: sbcl-slynk: Fix version number check.
* gnu/packages/lisp-xyz.scm (sbcl-slynk)[source]: Don't move files around
since this is not required.
[arguments]: Re-enable tests.  There is none at the moment, but they might get
added in the future.
Don't list systems since they are found automatically.

Moving `slynk.lisp' had the detrimental side-effect to break
`sly-version-string' which expects `sly.el' to be in the parent directory.  As
a result, using `sly-connect' from Emacs would prompt the user:

    [sly] Versions differ: 1.0.43 (sly) vs. nil (slynk). Continue? (y or n)

even though the Slynk version is correct.

Maybe there was a need for moving files around in the past, but since our ASDF
build system got revamped, all systems are now automatically found.  Thus we
can leave the files where they are, which fixes the version number check.