gnu: cheese: Add gst-plugins-good as an input.
This fixes an issue with choppy video recording. It also makes these lines not
appear in the build log:

meson.build:154: WARNING: vp8enc was not found. It needs to be installed before Cheese is run
meson.build:154: WARNING: webmmux was not found. It needs to be installed
before Cheese is run

* gnu/packages/gnome.scm (cheese)[inputs]: Add gst-plugins-good.