gnu: java-snappy: Fix checks when build machine has too much memory.
* gnu/packages/java-compression.scm (java-snappy)[arguments]: Add
set-test-memory-size phase.

This fixes <https://debbugs.gnu.org/cgi/bugreport.cgi?bug=46780>.
Thanks to zimoun for the solving idea.

The CachingBufferPoolTest tries to allocate a lot of memory (20GB) and the
test succeeds if there is not enough memory available in the JVM.
When the build machine has a lot of memory (which is the case on the Guix
CI system), the JVM will use more than 20GB as maximum heap size. In
consequence, the test fails.

The new phase adds a maxmemory size of 2G for the tests. Thus all other
tests will have enough memory and the CachingBufferPoolTest will fail.