gnu: Add java-jline-terminal.
This is part of JLine 3.

I was able to get this working properly under Guix by providing ncurses'
infocmp, as well as ensuring the *.caps files were present in the JAR,
but similar methods didn't work for the tests; if you have Java
knowledge, I'd appreciate the help getting those tests enabled.

* gnu/packages/java.scm (java-jline-terminal): New variable.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>