doc: Add a note about networking requirements for SSH access to VMs.
Fixes <https://issues.guix.gnu.org/48739>.  It may seem silly, but when not
realizing this firsthand, it can be quite surprising and thus not obvious to
troubleshoot.  Thanks to Eric Brown for their insightful comments on the
tracker.

* doc/guix.texi (Running Guix in a VM): Add a note that networking support
must be available in the Guix System operating system declaration used for the
VM.